# Various Gen refactors for better performance.