There are five changes which I've made to help ScalaCheck generate
values faster. They are:

 * Use custom buildableOfN implementation instead of sequence
 * Use Gen.stringOf and Gen.stringOfN instead of generic builders
 * Rewrite Gen[Char] instances to be faster
 * Remove as much indirection as possible
 * Stop using sieve/sieveCopy internally

There are also some changes that clean things up:

 * Add Gen.unicodeChar and Gen.unicodeStr
 * Add a Buildable[T, Seq[T]] instance
 * Add type annotations on some methods without them
 * A few indentation and formatting changes

The first 4 optimizations should not change any behavior users see
(except possibly balancing out some unbalanced character distributions
we had prevoiusly). The last change is a bit more controversial, and
will be discussed below.

Here are our benchmarks before the changes:

```
Benchmark                   (genSize)  (seedCount)  Mode  Cnt     Score      Error  Units
GenBench.asciiPrintableStr        100          100  avgt    3  1707.101 ± 2575.431  us/op
GenBench.const_                   100          100  avgt    3     3.431 ±    1.417  us/op
GenBench.double_                  100          100  avgt    3    13.103 ±   42.670  us/op
GenBench.dynamicFrequency         100          100  avgt    3  1804.406 ±  753.188  us/op
GenBench.eitherIntInt             100          100  avgt    3    42.499 ±   14.959  us/op
GenBench.identifier               100          100  avgt    3  3493.749 ± 1272.181  us/op
GenBench.int_                     100          100  avgt    3    13.364 ±    6.165  us/op
GenBench.listOfInt                100          100  avgt    3  1536.995 ±  370.880  us/op
GenBench.mapOfIntInt              100          100  avgt    3  3375.948 ±  588.872  us/op
GenBench.oneOf                    100          100  avgt    3    19.542 ±   14.573  us/op
GenBench.optionInt                100          100  avgt    3    52.750 ±    2.814  us/op
GenBench.sequence                 100          100  avgt    3   215.349 ±    0.736  us/op
GenBench.staticFrequency          100          100  avgt    3  1472.478 ±  655.671  us/op
GenBench.zipIntInt                100          100  avgt    3    26.897 ±    5.523  us/op
```

And after:

```
Benchmark                   (genSize)  (seedCount)  Mode  Cnt     Score      Error  Units
GenBench.asciiPrintableStr        100          100  avgt    3   204.613 ±  208.688  us/op
GenBench.const_                   100          100  avgt    3     2.856 ±    6.079  us/op
GenBench.double_                  100          100  avgt    3    11.058 ±   17.209  us/op
GenBench.dynamicFrequency         100          100  avgt    3   498.426 ±  306.193  us/op
GenBench.eitherIntInt             100          100  avgt    3    33.225 ±   10.462  us/op
GenBench.identifier               100          100  avgt    3    96.309 ±    4.903  us/op
GenBench.int_                     100          100  avgt    3     9.213 ±    2.990  us/op
GenBench.listOfInt                100          100  avgt    3   448.545 ±  228.448  us/op
GenBench.mapOfIntInt              100          100  avgt    3  1778.107 ± 1532.473  us/op
GenBench.oneOf                    100          100  avgt    3    18.858 ±   24.069  us/op
GenBench.optionInt                100          100  avgt    3    45.695 ±   18.507  us/op
GenBench.sequence                 100          100  avgt    3   198.399 ±  164.050  us/op
GenBench.staticFrequency          100          100  avgt    3   449.571 ±   97.772  us/op
GenBench.zipIntInt                100          100  avgt    3    24.025 ±   27.260  us/op
```

Some notable speed increases include:

 * asciiPrintableStr (8.5x faster)
 * dynamicFrequency (3.6x faster)
 * identifier (36.4x faster)
 * listOfInt (3.4x faster)
 * mapOfIntInt (1.9x faster)
 * staticFrequency (3.3x faster)

The speed improvements mostly affects collections, particularly
strings. Since these represent a significant portion of values
ScalaCheck users are likely to generate, this will probably help
shorten test runtime for our users. For example, running the Paiges
tests in this branch resulted in a roughly 2x speed up in time (as
measured by SBT).

The one optimization we might chose to forgoe is deprecating sieves.
Sieves were introduced to allow filtering predicates (introduced by
calling .suchThat or .filter on Gen[T] values) to be preserved in the
generated results. One strange consequence of that is that Gen.R holds
onto values which might be filtered out -- the acutal filtering is
done when .retrieve is called. Another is that many of the collection
combinators include .forall checks to try to verify that their
elements are legitimate.

The reason sieves were added was to support shrinking. The shrinking
code uses the sieve to filter the stream of shrunken values to try to
ensure that only "valid" values are produced. Since users tend to
avoid using filter (because of the issues around flaky test failures
due to too many discarded cases) and since most actual Gen instances
fail to shrink properly anyway, these sieves have likely not benefited
too many users. However, there's a risk that for some users removing
sieves will exacerbate shrinking issues they already have.

I'd like to consider removing sieves, since as it stands I'm not sure
they are consistently used, and using them doesn't solve the problems
we have with shrinking. However, I'll also benchmark this branch with
sieves put back in, and if the difference in performance is minor we
may want to leave sieves alone for now.