@non 2019-09-28
@johnynek Thanks for the thorough review! I just pushed some code that addresses many of your comments, but not all of them. I do seem to  have made some combinators a bit slower so I want to consider how to proceed next.