@non 2019-09-28
Apparently changing `lazy val` to `val` breaks binary compatibility. Who knew?