@ashawley 2019-09-29
I ran the benchmark for the fist commit, 126168f.

Before (current master):
```
GenBench.arbitraryString    avgt    3   4191.12 ± 14057.61  us/op
GenBench.asciiPrintableStr  avgt    3   2254.36 ±  2951.37  us/op
GenBench.const_             avgt    3      4.39 ±     0.60  us/op
GenBench.double_            avgt    3     13.53 ±     3.97  us/op
GenBench.dynamicFrequency   avgt    3   2159.29 ±   347.15  us/op
GenBench.eitherIntInt       avgt    3     58.70 ±    12.89  us/op
GenBench.identifier         avgt    3   4862.64 ±   493.61  us/op
GenBench.int_               avgt    3     17.55 ±     2.89  us/op
GenBench.listOfInt          avgt    3   2008.92 ±   561.94  us/op
GenBench.mapOfIntInt        avgt    3   4607.00 ±   687.68  us/op
GenBench.oneOf              avgt    3     26.04 ±     2.70  us/op
GenBench.optionInt          avgt    3     95.90 ±   440.56  us/op
GenBench.sequence           avgt    3    324.61 ±    77.22  us/op
GenBench.staticFrequency    avgt    3   2081.64 ±   536.27  us/op
GenBench.testFilter         avgt    3  23920.79 ± 10953.43  us/op
GenBench.zipIntInt          avgt    3     37.93 ±     0.85  us/op
```
After (126168f):
```
GenBench.arbitraryString    avgt    3    438.94 ±   539.99  us/op
GenBench.asciiPrintableStr  avgt    3    316.78 ±   568.60  us/op
GenBench.const_             avgt    3      3.05 ±     0.24  us/op
GenBench.double_            avgt    3     13.20 ±     8.80  us/op
GenBench.dynamicFrequency   avgt    3    748.58 ±   702.64  us/op
GenBench.eitherIntInt       avgt    3     48.74 ±    41.64  us/op
GenBench.identifier         avgt    3    224.31 ±    89.49  us/op
GenBench.int_               avgt    3     14.27 ±     0.45  us/op
GenBench.listOfInt          avgt    3    623.88 ±    20.94  us/op
GenBench.mapOfIntInt        avgt    3   2536.78 ±   117.44  us/op
GenBench.oneOf              avgt    3     24.77 ±    13.41  us/op
GenBench.optionInt          avgt    3     65.45 ±    32.17  us/op
GenBench.sequence           avgt    3    273.03 ±    17.93  us/op
GenBench.staticFrequency    avgt    3    654.64 ±   130.07  us/op
GenBench.testFilter         avgt    3   1424.34 ±   282.77  us/op
GenBench.zipIntInt          avgt    3     34.77 ±    33.82  us/op
```

Confirms the improvements Erik observed above, different only from some variation:

- arbitraryString (9.5x faster)
- asciiPrintableStr (7x faster)
- dynamicFrequency (3x faster)
- identifier (21x faster)
- listOfInt (3.2x faster)
- mapOfIntInt (1.8x faster)
- staticFrequency (3.2x faster)