@non 2019-09-30
The outstanding work on this branch is:

1. Try restoring the sieves and see how much of the improvement we lose, if any.
2. Consider whether removing sieves is on the table or not.
3. Think about a taking a more principled stance on retries.
4. Try doing some more profiling to see if there's anywhere else we can save a lot of time.

I'd also like to try out this branch on more other repositories to see what the total wall clock time improvement using this branch is.