@ashawley 2020-01-08
I'm going to work on cherry-picking this apart so that it can get closer to crossing the finish line.  I noticed a lot of the improvements here are not very controversial.  Isolating them out and merging them will make that case clearer.

This exercise also helps me study the changes.  I don't have good intuition about this aspect of the library. Erik is likely the singular expert of the seed implementation that he contributed to ScalaCheck.

Merging the changes separately will also make it easier to troubleshoot any potential regressions.  Any bugs are most likely from the challenge of improving an old but popular property-testing library, not because of sloppy coding.  Since the Travis build publishes snapshots, it's also worth going this route of merging the individual changes, should we need to verify the changes.