@larsrh 2020-10-26
I think all of these changes have now been incorporated. Feel free to reopen if this isn't the case.